requirements.txt
Django==1.9.6
Pillow==3.2.0
Unidecode==0.04.19
django-debug-toolbar==1.4
django-filter==0.13.0
django-simple-captcha==0.5.1
djangorestframework==3.3.3
pytz==2016.4
six==1.10.0
sqlparse==0.1.19
